move DispatchWorkerType out of API directory
