+++ tt_metal/api/tt-metalium/dispatch_core_common.hpp
@@ -13,21 +13,6 @@
 
 namespace tt::tt_metal {
 
-enum DispatchWorkerType : uint32_t {
-    PREFETCH = 0,
-    PREFETCH_HD = 1,
-    PREFETCH_H = 2,
-    PREFETCH_D = 3,
-    DISPATCH = 4,
-    DISPATCH_HD = 5,
-    DISPATCH_H = 6,
-    DISPATCH_D = 7,
-    DISPATCH_S = 8,
-    FABRIC_MUX = 17,         // Downstream from MMIO to remote mux. Tunnel index is required.
-    RETURN_FABRIC_MUX = 18,  // Upstream from remote to MMIO mux. Tunnel index will be determined from the device id.
-    COUNT,
-};
-
 enum class DispatchCoreType : uint32_t { WORKER, ETH, COUNT };
 
 enum class DispatchCoreAxis { ROW, COL, COUNT };
