+++ tt_metal/impl/dispatch/dispatch_core_common.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "dispatch_core_common.hpp"
+#include "impl/dispatch/dispatch_core_common.hpp"
 #include "impl/context/metal_context.hpp"
 #include <umd/device/types/arch.hpp>
 #include <umd/device/types/core_coordinates.hpp>
@@ -23,9 +23,5 @@ DispatchCoreAxis DispatchCoreConfig::get_default_axis() {
 DispatchCoreConfig get_dispatch_core_config() {
     return MetalContext::instance().get_dispatch_core_manager().get_dispatch_core_config();
 }
-
-CoreType get_dispatch_core_type() {
-    return MetalContext::instance().get_dispatch_core_manager().get_dispatch_core_type();
-};
 
 }  // namespace tt::tt_metal
